requirements.txt
Flask
pymongo
dnspython
python-dotenv
Flask-PyMongo
Flask-JWT-Extended
flask_pymongo
PyJWT
mongoengine
flask_jwt_extended
